Use ensure_started implementation to implement split
