+++ libs/parallelism/execution/tests/unit/algorithm_split.cpp
@@ -21,8 +21,8 @@ namespace ex = hpx::execution::experimental;
 // This overload is only used to check dispatching. It is not a useful
 // implementation.
 template <typename Allocator = hpx::util::internal_allocator<>>
-auto tag_dispatch(ex::ensure_started_t, custom_sender_tag_dispatch s,
-    Allocator const& = Allocator{})
+auto tag_dispatch(
+    ex::split_t, custom_sender_tag_dispatch s, Allocator const& = Allocator{})
 {
     s.tag_dispatch_overload_called = true;
     return void_sender{};
@@ -33,10 +33,8 @@ int main()
     // Success path
     {
         std::atomic<bool> set_value_called{false};
-        std::atomic<bool> started{false};
-        auto s1 = ex::transform(void_sender{}, [&]() { started = true; });
-        auto s2 = ex::ensure_started(std::move(s1));
-        HPX_TEST(started);
+        auto s1 = void_sender{};
+        auto s2 = ex::split(std::move(s1));
         auto f = [] {};
         auto r = callback_receiver<decltype(f)>{f, set_value_called};
         auto os = ex::connect(std::move(s2), std::move(r));
@@ -46,13 +44,8 @@ int main()
 
     {
         std::atomic<bool> set_value_called{false};
-        std::atomic<bool> started{false};
-        auto s1 = ex::transform(ex::just(0), [&](int x) {
-            started = true;
-            return x;
-        });
-        auto s2 = ex::ensure_started(std::move(s1));
-        HPX_TEST(started);
+        auto s1 = ex::just(0);
+        auto s2 = ex::split(std::move(s1));
         auto f = [](int x) { HPX_TEST_EQ(x, 0); };
         auto r = callback_receiver<decltype(f)>{f, set_value_called};
         auto os = ex::connect(std::move(s2), std::move(r));
@@ -62,15 +55,8 @@ int main()
 
     {
         std::atomic<bool> set_value_called{false};
-        std::atomic<bool> started{false};
-        auto s1 =
-            ex::transform(ex::just(custom_type_non_default_constructible{42}),
-                [&](custom_type_non_default_constructible x) {
-                    started = true;
-                    return x;
-                });
-        auto s2 = ex::ensure_started(std::move(s1));
-        HPX_TEST(started);
+        auto s1 = ex::just(custom_type_non_default_constructible{42});
+        auto s2 = ex::split(std::move(s1));
         auto f = [](auto x) { HPX_TEST_EQ(x.x, 42); };
         auto r = callback_receiver<decltype(f)>{f, set_value_called};
         auto os = ex::connect(std::move(s2), std::move(r));
@@ -80,15 +66,9 @@ int main()
 
     {
         std::atomic<bool> set_value_called{false};
-        std::atomic<bool> started{false};
-        auto s1 = ex::transform(
-            ex::just(custom_type_non_default_constructible_non_copyable{42}),
-            [&](custom_type_non_default_constructible_non_copyable&& x) {
-                started = true;
-                return std::move(x);
-            });
-        auto s2 = ex::ensure_started(std::move(s1));
-        HPX_TEST(started);
+        auto s1 =
+            ex::just(custom_type_non_default_constructible_non_copyable{42});
+        auto s2 = ex::split(std::move(s1));
         auto f = [](auto& x) { HPX_TEST_EQ(x.x, 42); };
         auto r = callback_receiver<decltype(f)>{f, set_value_called};
         auto os = ex::connect(std::move(s2), std::move(r));
@@ -99,7 +79,7 @@ int main()
     // operator| overload
     {
         std::atomic<bool> set_value_called{false};
-        auto s = void_sender{} | ex::ensure_started();
+        auto s = void_sender{} | ex::split();
         auto f = [] {};
         auto r = callback_receiver<decltype(f)>{f, set_value_called};
         auto os = ex::connect(std::move(s), std::move(r));
@@ -112,7 +92,7 @@ int main()
         std::atomic<bool> receiver_set_value_called{false};
         std::atomic<bool> tag_dispatch_overload_called{false};
         auto s = custom_sender_tag_dispatch{tag_dispatch_overload_called} |
-            ex::ensure_started();
+            ex::split();
         auto f = [] {};
         auto r = callback_receiver<decltype(f)>{f, receiver_set_value_called};
         auto os = ex::connect(std::move(s), std::move(r));
@@ -124,7 +104,7 @@ int main()
     // Failure path
     {
         std::atomic<bool> set_error_called{false};
-        auto s = error_sender{} | ex::ensure_started();
+        auto s = error_sender{} | ex::split();
         auto r = error_callback_receiver<decltype(check_exception_ptr)>{
             check_exception_ptr, set_error_called};
         auto os = ex::connect(std::move(s), std::move(r));
@@ -134,8 +114,7 @@ int main()
 
     {
         std::atomic<bool> set_error_called{false};
-        auto s = error_sender{} | ex::ensure_started() | ex::ensure_started() |
-            ex::ensure_started();
+        auto s = error_sender{} | ex::split() | ex::split() | ex::split();
         auto r = error_callback_receiver<decltype(check_exception_ptr)>{
             check_exception_ptr, set_error_called};
         auto os = ex::connect(std::move(s), std::move(r));
@@ -143,13 +122,13 @@ int main()
         HPX_TEST(set_error_called);
     }
 
-    // Chained ensure_started calls do not create new shared states
+    // Chained split calls do not create new shared states
     {
         std::atomic<bool> receiver_set_value_called{false};
-        auto s1 = ex::just() | ex::ensure_started();
-        auto s2 = ex::ensure_started(s1);
+        auto s1 = ex::just() | ex::split();
+        auto s2 = ex::split(s1);
         HPX_TEST_EQ(s1.state, s2.state);
-        auto s3 = ex::ensure_started(std::move(s2));
+        auto s3 = ex::split(std::move(s2));
         HPX_TEST_EQ(s1.state, s3.state);
         auto f = [] {};
         auto r = callback_receiver<decltype(f)>{f, receiver_set_value_called};
@@ -160,10 +139,10 @@ int main()
 
     {
         std::atomic<bool> receiver_set_value_called{false};
-        auto s1 = ex::just(42) | ex::ensure_started();
-        auto s2 = ex::ensure_started(s1);
+        auto s1 = ex::just(42) | ex::split();
+        auto s2 = ex::split(s1);
         HPX_TEST_EQ(s1.state, s2.state);
-        auto s3 = ex::ensure_started(std::move(s2));
+        auto s3 = ex::split(std::move(s2));
         HPX_TEST_EQ(s1.state, s3.state);
         auto f = [](int x) { HPX_TEST_EQ(x, 42); };
         auto r = callback_receiver<decltype(f)>{f, receiver_set_value_called};
